:fix: logic optimize input image in main.py and fix logic AI entity, service, repo

+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/AiAnalysisDto.java
@@ -1,14 +1,18 @@
 package com.smartfarm.api.dto;
 
-import lombok.*;
 import java.time.LocalDateTime;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class AiAnalysisDto {
-    private Long analysisId;       // bigint -> giữ Long
+    private Integer analysisId;
     private Integer pBatchId;
     private String batchName;
     private String imagePath;
