+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/service/AiAnalysisService.java
@@ -62,7 +62,7 @@ public class AiAnalysisService {
         return aiAnalysisRepository.findByPlantingBatchPBatchId(pBatchId).stream().map(aiAnalysisMapper::toDto).collect(Collectors.toList());
     }
 
-    public Optional<AiAnalysisDto> findById(Long id) {
+    public Optional<AiAnalysisDto> findById(Integer id) {
         return aiAnalysisRepository.findById(id).map(aiAnalysisMapper::toDto);
     }
 
@@ -71,14 +71,14 @@ public class AiAnalysisService {
         return aiAnalysisMapper.toDto(aiAnalysisRepository.save(entity));
     }
 
-    public Optional<AiAnalysisDto> update(Long id, AiAnalysisDto dto) {
+    public Optional<AiAnalysisDto> update(Integer id, AiAnalysisDto dto) {
         if (!aiAnalysisRepository.existsById(id)) return Optional.empty();
         AiAnalysis entity = aiAnalysisMapper.toEntity(dto);
         entity.setAnalysisId(id);
         return Optional.of(aiAnalysisMapper.toDto(aiAnalysisRepository.save(entity)));
     }
 
-    public boolean deleteById(Long id) {
+    public boolean deleteById(Integer id) {
         if (!aiAnalysisRepository.existsById(id)) return false;
         aiAnalysisRepository.deleteById(id);
         return true;
